Bts 101/category read

+++ src/main/java/com/readme/sections/dto/NovelCardsDTO.java
@@ -28,7 +28,7 @@ public class NovelCardsDTO {
     String serializationStatus;
     List<Tag> tags;
     Long scheduleId;
-    Float starRating;
+    Double starRating;
     Boolean monday;
     Boolean tuesday;
     Boolean wednesday;
